+++ jwst/extract_1d/extract.py
@@ -1011,7 +1011,7 @@ def create_poly(coeff):
         key = "c{}".format(i)
         coeff_dict[key] = coeff[i]
 
-    return polynomial.Polynomial1D(degree=n - 1, **coeff_dict)
+    return polynomial.Polynomial1D(degree=n-1, **coeff_dict)
 
 
 class ExtractBase:
@@ -1299,7 +1299,7 @@ class ExtractBase:
 class ExtractModel(ExtractBase):
     """The extraction region was specified in a JSON file."""
 
-    def __init__(self, input_model, slit,
+    def __init__(self, input_model, slit, verbose,
                  ref_file_type=None,
                  match="unknown",
                  dispaxis=HORIZONTAL, spectral_order=1,
@@ -1308,7 +1308,8 @@ class ExtractModel(ExtractBase):
                  independent_var="pixel",
                  smoothing_length=0, bkg_order=0, nod_correction=0.,
                  x_center=None, y_center=None,
-                 inner_bkg=None, outer_bkg=None, method='subpixel'):
+                 inner_bkg=None, outer_bkg=None, method='subpixel',
+                 subtract_background=None):
         """Create a polynomial model from coefficients.
 
         Extended summary
@@ -1400,6 +1401,12 @@ class ExtractModel(ExtractBase):
 
         method : str
             This is not relevant; it's only used for IFU data.
+
+        subtract_background : bool or None
+            A flag which indicates whether the background should be subtracted.
+            If None, the value in the extract_1d reference file will be used.
+            If not None, this parameter overrides the value in the
+            extract_1d reference file.
         """
 
         super().__init__()
@@ -1488,6 +1495,16 @@ class ExtractModel(ExtractBase):
         # corresponding polynomial functions.
         self.src_coeff = copy.deepcopy(src_coeff)
         self.bkg_coeff = copy.deepcopy(bkg_coeff)
+
+        self.subtract_background = subtract_background
+        if not subtract_background:
+            if verbose and self.bkg_coeff is not None:
+                log.info("Background subtraction was turned off - skipping it.")
+            self.bkg_coeff = None
+        else:
+            if verbose and self.bkg_coeff is None:
+                log.info("Skipping background subtraction because "
+                         "background regions are not defined.")
 
         # These functions will be assigned by assign_polynomial_limits.
         # The "p" in the attribute name indicates a polynomial function.
@@ -1750,17 +1767,20 @@ class ExtractModel(ExtractBase):
             ra and dec are the right ascension and declination respectively
             at the nominal center of the slit.
 
-        wavelength : ndarray, 1-D
+        wavelength : ndarray, 1-D, float64
             The wavelength in micrometers at each pixel.
 
         net : ndarray, 1-D
             The count rate (counts / s) minus the background at each pixel.
 
-        background : ndarray, 1-D
+        background : ndarray, 1-D, float64
             The background count rate that was subtracted from the total
             source count rate to get `net`.
 
-        dq : ndarray, 1-D, int32
+        npixels : ndarray, 1-D, float64
+            The number of pixels that were added together to get `net`.
+
+        dq : ndarray, 1-D, uint32
             The data quality array.
         """
 
@@ -1902,19 +1922,20 @@ class ExtractModel(ExtractBase):
             temp_wl[nan_mask] = 0.01            # because NaNs cause problems
 
         # src total flux, area, total weight
-        (net, background) = \
+        (net, background, npixels) = \
         extract1d.extract1d(image, temp_wl, disp_range,
                             self.p_src, self.p_bkg, self.independent_var,
                             self.smoothing_length, self.bkg_order,
-                            weights=None)
+                            weights=None, subtract_background=self.subtract_background)
         del temp_wl
 
-        dq = np.zeros(net.shape, dtype=np.int32)
+        dq = np.zeros(net.shape, dtype=np.uint32)
         if n_nan > 0:
-            (wavelength, net, background, dq) = \
-                nans_at_endpoints(wavelength, net, background, dq, verbose)
+            (wavelength, net, background, npixels, dq) = \
+                nans_at_endpoints(wavelength, net, background, npixels, dq,
+                                  verbose)
 
-        return (ra, dec, wavelength, net, background, dq)
+        return (ra, dec, wavelength, net, background, npixels, dq)
 
 
 class ImageExtractModel(ExtractBase):
@@ -1927,7 +1948,8 @@ class ImageExtractModel(ExtractBase):
                  ref_image=None,
                  dispaxis=HORIZONTAL,
                  smoothing_length=0,
-                 nod_correction=0):
+                 nod_correction=0,
+                 subtract_background=None):
         """Extract using a reference image to define the extraction and
            background regions.
 
@@ -1998,6 +2020,7 @@ class ImageExtractModel(ExtractBase):
                         smoothing_length)
             smoothing_length += 1               # must be odd
         self.smoothing_length = smoothing_length
+        self.subtract_background = subtract_background
 
         if self.exp_type == "NIS_SOSS":
             if hasattr(input_model.meta, 'wcs'):
@@ -2110,7 +2133,10 @@ class ImageExtractModel(ExtractBase):
             The background count rate that was subtracted from the total
             source count rate to get `net`.
 
-        dq : ndarray, 1-D, int32
+        npixels : ndarray, 1-D, float32
+            The number of pixels that were added together to get `net`.
+
+        dq : ndarray, 1-D, uint32
         """
 
         shape = data.shape
@@ -2134,6 +2160,18 @@ class ImageExtractModel(ExtractBase):
         # Extract the data.
         gross = (data * mask_target).sum(axis=axis, dtype=np.float)
 
+        # Compute the number of pixels that were added together to get gross.
+        temp = np.ones_like(data)
+        npixels = (temp * mask_target).sum(axis=axis, dtype=np.float)
+
+        if not self.subtract_background:
+            if verbose and mask_bkg is not None:
+                    log.info("Background subtraction was turned off - skipping it.")
+            mask_bkg = None
+        else:
+            if verbose and mask_bkg is None:
+                    log.info("Skipping background subtraction because "
+                             "background regions are not defined.")
         # Extract the background.
         if mask_bkg is not None:
             n_bkg = mask_bkg.sum(axis=axis, dtype=np.float)
@@ -2298,16 +2336,17 @@ class ImageExtractModel(ExtractBase):
                 wavelength = np.arange(shape[0], dtype=np.float)
             wavelength = wavelength[trim_slc]
 
-        dq = np.zeros(net.shape, dtype=np.int32)
+        dq = np.zeros(net.shape, dtype=np.uint32)
         nan_mask = np.isnan(wavelength)
         n_nan = nan_mask.sum(dtype=np.intp)
         if n_nan > 0:
             if verbose:
                 log.warning("%d NaNs in wavelength array", n_nan)
-            (wavelength, net, background, dq) = \
-                nans_at_endpoints(wavelength, net, background, dq, verbose)
+            (wavelength, net, background, npixels, dq) = \
+                nans_at_endpoints(wavelength, net, background, npixels, dq,
+                                  verbose)
 
-        return (ra, dec, wavelength, net, background, dq)
+        return (ra, dec, wavelength, net, background, npixels, dq)
 
 
     def match_shape(self, shape):
@@ -2455,7 +2494,7 @@ def interpolate_response(wavelength, relsens, verbose):
 
 
 def do_extract1d(input_model, refname, smoothing_length, bkg_order,
-                 log_increment):
+                 log_increment, subtract_background):
     """Extract 1-D spectra.
 
     Extended summary
@@ -2481,6 +2520,12 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
         multi-integration, a message will be written to the log every
         `log_increment` integrations.
 
+    subtract_background : bool or None
+        User supplied flag indicating whether the background should be subtracted.
+        If None, the value in the extract_1d reference file will be used.
+        If not None, this parameter overrides the value in the
+        extract_1d reference file.
+
     Returns
     -------
     output_model : data model
@@ -2491,6 +2536,9 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
     if hasattr(input_model, "int_times"):
         output_model.int_times = input_model.int_times.copy()
     output_model.update(input_model)
+
+    # This data type is used for creating an output table.
+    spec_dtype = datamodels.SpecModel().spec_table.dtype
 
     # This will be relevant if we're asked to extract a spectrum and the
     # spectral order is zero.  That's only OK if the disperser is a prism.
@@ -2522,6 +2570,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                 ref_dict,
                                 slit, slit.name, sp_order,
                                 input_model.meta, smoothing_length, bkg_order)
+            if subtract_background is not None:
+                extract_params['subtract_background'] = subtract_background
             if extract_params['match'] == NO_MATCH:
                 log.critical('Missing extraction parameters.')
                 raise ValueError('Missing extraction parameters.')
@@ -2535,7 +2585,7 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 continue
 
             try:
-                (ra, dec, wavelength, net, background, dq,
+                (ra, dec, wavelength, net, background, npixels, dq,
                  prev_offset) = extract_one_slit(
                                         input_model, slit, -1,
                                         prev_offset, True, extract_params)
@@ -2559,10 +2609,10 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
             fl_error = np.ones_like(net)
             nerror = np.ones_like(net)
             berror = np.ones_like(net)
-            spec = datamodels.SpecModel()
             otab = np.array(list(zip(wavelength, flux, fl_error, dq,
-                                 net, nerror, background, berror)),
-                            dtype=spec.spec_table.dtype)
+                                     net, nerror, background, berror,
+                                     npixels)),
+                            dtype=spec_dtype)
             spec = datamodels.SpecModel(spec_table=otab)
             spec.meta.wcs = spec_wcs.create_spectral_wcs(ra, dec, wavelength)
             spec.spec_table.columns['wavelength'].unit = 'um'
@@ -2610,6 +2660,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
                                     bkg_order)
+                if subtract_background is not None:
+                    extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == EXACT:
                     slit = DUMMY
                     find_dispaxis(input_model, slit, sp_order, extract_params)
@@ -2618,7 +2670,7 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                     "determined, so skipping ...")
                         continue
                     try:
-                        (ra, dec, wavelength, net, background, dq,
+                        (ra, dec, wavelength, net, background, npixels, dq,
                          prev_offset) = extract_one_slit(
                                         input_model, slit, -1,
                                         prev_offset, True, extract_params)
@@ -2649,10 +2701,10 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 fl_error = np.ones_like(net)
                 nerror = np.ones_like(net)
                 berror = np.ones_like(net)
-                spec = datamodels.SpecModel()
                 otab = np.array(list(zip(wavelength, flux, fl_error, dq,
-                                     net, nerror, background, berror)),
-                                dtype=spec.spec_table.dtype)
+                                         net, nerror, background, berror,
+                                         npixels)),
+                                dtype=spec_dtype)
                 spec = datamodels.SpecModel(spec_table=otab)
                 spec.meta.wcs = spec_wcs.create_spectral_wcs(
                                         ra, dec, wavelength)
@@ -2690,6 +2742,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
                                     bkg_order)
+                if subtract_background is not None:
+                    extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == NO_MATCH:
                     log.critical('Missing extraction parameters.')
                     raise ValueError('Missing extraction parameters.')
@@ -2724,7 +2778,7 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 for integ in range(input_model.data.shape[0]):
                     # Extract spectrum
                     try:
-                        (ra, dec, wavelength, net, background, dq,
+                        (ra, dec, wavelength, net, background, npixels, dq,
                          prev_offset) = extract_one_slit(
                                         input_model, slit, integ,
                                         prev_offset, verbose, extract_params)
@@ -2741,10 +2795,10 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                     fl_error = np.ones_like(net)
                     nerror = np.ones_like(net)
                     berror = np.ones_like(net)
-                    spec = datamodels.SpecModel()
                     otab = np.array(list(zip(wavelength, flux, fl_error, dq,
-                                         net, nerror, background, berror)),
-                                    dtype=spec.spec_table.dtype)
+                                             net, nerror, background, berror,
+                                             npixels)),
+                                             dtype=spec_dtype)
                     spec = datamodels.SpecModel(spec_table=otab)
                     spec.meta.wcs = spec_wcs.create_spectral_wcs(
                                         ra, dec, wavelength)
@@ -2790,7 +2844,7 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 source_type = input_model.meta.target.source_type.lower()
             except AttributeError:
                 source_type = "unknown"
-            output_model = ifu.ifu_extract1d(input_model, refname, source_type)
+            output_model = ifu.ifu_extract1d(input_model, refname, source_type, subtract_background)
 
         else:
             log.error("The input file is not supported for this step.")
@@ -2989,7 +3043,7 @@ def is_prism(input_model):
     Extended summary
     ----------------
     The reason for this test is so we can skip spectral extraction if the
-    spectral order is zero and the exposure was not made using a prism.  
+    spectral order is zero and the exposure was not made using a prism.
     In this context, therefore, a grism is not considered to be a prism.
 
     Parameters
@@ -3111,17 +3165,20 @@ def extract_one_slit(input_model, slit, integ,
         ra and dec are the right ascension and declination respectively
         at the nominal center of the slit.
 
-    wavelength : ndarray, 1-D
+    wavelength : ndarray, 1-D, float64
         The wavelength in micrometers at each pixel.
 
-    net : ndarray, 1-D
+    net : ndarray, 1-D, float64
         The count rate (counts / s) minus the background at each pixel.
 
-    background : ndarray, 1-D
+    background : ndarray, 1-D, float64
         The background count rate that was subtracted from the total
         source count rate to get `net`.
 
-    dq : ndarray, 1-D, int32
+    npixels : ndarray, 1-D, float64
+        The number of pixels that were added together to get `net`.
+
+    dq : ndarray, 1-D, uint32
         The data quality array.
 
     offset : float
@@ -3163,12 +3220,12 @@ def extract_one_slit(input_model, slit, integ,
 
     if extract_params['ref_file_type'] == FILE_TYPE_IMAGE:
         # The reference file is an image.
-        extract_model = ImageExtractModel(input_model, slit, **extract_params)
+        extract_model = ImageExtractModel(input_model, slit, verbose, **extract_params)
         ap = None
     else:
         # If there is a reference file (there doesn't have to be), it's in
         # JSON format.
-        extract_model = ExtractModel(input_model, slit, **extract_params)
+        extract_model = ExtractModel(input_model, slit, verbose, **extract_params)
         ap = get_aperture(data.shape, extract_model.wcs,
                           verbose, extract_params)
         extract_model.update_extraction_limits(ap)
@@ -3193,10 +3250,10 @@ def extract_one_slit(input_model, slit, integ,
         extract_model.log_extraction_parameters()
 
     extract_model.assign_polynomial_limits(verbose)
-    (ra, dec, wavelength, net, background, dq) = \
+    (ra, dec, wavelength, net, background, npixels, dq) = \
                 extract_model.extract(data, wl_array, verbose)
 
-    return (ra, dec, wavelength, net, background, dq, offset)
+    return (ra, dec, wavelength, net, background, npixels, dq, offset)
 
 
 def replace_bad_values(data, input_dq, fill=0.):
@@ -3232,7 +3289,7 @@ def replace_bad_values(data, input_dq, fill=0.):
     else:
         return data
 
-def nans_at_endpoints(wavelength, net, background, dq, verbose):
+def nans_at_endpoints(wavelength, net, background, npixels, dq, verbose):
     """Flag NaNs in the wavelength array.
 
     Extended summary
@@ -3254,6 +3311,9 @@ def nans_at_endpoints(wavelength, net, background, dq, verbose):
     background : ndarray
         Array of background values that were subtracted to get `net`.
 
+    npixels : ndarray, float32
+        The number of pixels that were added together to get `net`.
+
     dq : ndarray
         Data quality array.
 
@@ -3262,15 +3322,16 @@ def nans_at_endpoints(wavelength, net, background, dq, verbose):
 
     Returns
     -------
-    wavelength, net, background, dq : ndarray
+    wavelength, net, background, npixels, dq : ndarray
         The returned `dq` array may have NaNs flagged with DO_NOT_USE,
-        and all four arrays may have been trimmed at either or both ends.
+        and all five arrays may have been trimmed at either or both ends.
     """
 
     # The input arrays will not be modified in-place.
     new_wl = wavelength.copy()
     new_net = net.copy()
     new_bkg = background.copy()
+    new_npixels = npixels.copy()
     new_dq = dq.copy()
     nelem = wavelength.shape[0]
 
@@ -3290,8 +3351,9 @@ def nans_at_endpoints(wavelength, net, background, dq, verbose):
             new_wl = new_wl[slc]
             new_net = new_net[slc]
             new_bkg = new_bkg[slc]
+            new_npixels = new_npixels[slc]
             new_dq = new_dq[slc]
     else:
         new_dq |= dqflags.pixel['DO_NOT_USE']
 
-    return (new_wl, new_net, new_bkg, new_dq)
+    return (new_wl, new_net, new_bkg, new_npixels, new_dq)