+++ jwst/master_background/master_background_step.py
@@ -1,5 +1,9 @@
+from os.path import basename
+
 from ..stpipe import Step
 from .. import datamodels
+
+from .expand_to_2d import expand_to_2d
 
 
 __all__ = ["MasterBackgroundStep"]
@@ -39,6 +43,9 @@ class MasterBackgroundStep(Step):
             The background-subtracted target data model(s)
         """
 
+        # Get association info if available
+        # asn_id = ???
+
         with datamodels.open(input) as input_data:
 
             # Handle individual NIRSpec FS, NIRSpec MOS
@@ -58,39 +65,90 @@ class MasterBackgroundStep(Step):
                 pass
 
             else:
-                result = input_data
+                result = input_data.copy()
                 self.log.warning(
                     "Input %s of type %s cannot be handled.  Step skipped.",
                     input, type(input)
                     )
-                try:
-                    result.meta.cal_step.master_back_sub = 'SKIPPED'
-                except AttributeError:
-                    for model in result:
-                        model.meta.cal_step.master_back_sub = 'SKIPPED'
+                self.record_step_status(result, 'master_background', success=False)
 
                 return result
 
             # Check if user has supplied a master background spectrum.
             if self.user_background is None:
-                pass
+                # TODO: 1. compute master background from asn, 2. subtract it
+                # Return input as dummy result for now
+                result = input_data.copy()
             else:
-                pass
+                background_2d = expand_to_2d(input_data, self.user_background)
+                result = subtract_2d_background(input_data, background_2d)
+
+                # Record name of user-supplied master background spectrum
+                if isinstance(result, datamodels.ModelContainer):
+                    for model in result:
+                        model.meta.background.master_background_file = basename(self.user_background)
+                else:
+                    result.meta.background.master_background_file = basename(self.user_background)
 
             # Save the computed background if requested by user
             if self.save_background and self.user_background is None:
+                # self.save_model(background, suffix='masterbg', asn_id=asn_id)
                 pass
-                # TODO, figure out step suffix, use stpipe tools to generate
-                # the filename from the input filename(s) and/or association.
-                # Notice this is a custom `output_file` type in the spec above
-
-            # Return input as dummy result for now
-            result = input_data
             
-            try:
-                result.meta.cal_step.master_back_sub = 'COMPLETE'
-            except AttributeError:
-                for model in result:
-                    model.meta.cal_step.master_back_sub = 'COMPLETE'
+            self.record_step_status(result, 'master_background')
 
         return result
+
+
+def subtract_2d_background(source, background):
+    """Subtract a 2D background
+
+    Parameters
+    ----------
+    source : `~jwst.datamodels.DataModel` or `~jwst.datamodels.ModelContainer`
+        The input science data.
+
+    background : `~jwst.datamodels.DataModel`
+        The input background data.  Must be the same datamodel type as `source`.
+        For a `~jwst.datamodels.ModelContainer`, the source and background
+        models in the input containers must match one-to-one.
+
+    Returns
+    -------
+    `~jwst.datamodels.DataModel`
+        Background subtracted from source.
+    """
+    def _subtract_2d_background(model, background):
+        result = model.copy()
+        # Handle individual NIRSpec FS, NIRSpec MOS
+        if isinstance(model, datamodels.MultiSlitModel):
+            for slit, slitbg in zip(result.slits, background.slits):
+                slit.data -= slitbg.data
+
+        # Handle MIRI LRS, MIRI MRS and NIRSpec IFU
+        elif isinstance(model, (datamodels.ImageModel, datamodels.IFUImageModel)):
+            result.data -= background.data
+
+        else:
+            # Shouldn't get here.
+            raise RuntimeError("Input type {} is not supported."
+                               .format(type(model)))
+        return result
+
+    # Handle containers of many datamodels
+    if isinstance(source, datamodels.ModelContainer):
+        result = datamodels.ModelContainer()
+        result.update(source)
+        for model, bg in zip(source, background):
+            result.append(_subtract_2d_background(model, bg))
+
+    # Handle single datamodels
+    elif isinstance(source, (datamodels.ImageModel, datamodels.IFUImageModel, datamodels.MultiSlitModel)):
+        result = _subtract_2d_background(source, background)
+
+    else:
+        # Shouldn't get here.
+        raise RuntimeError("Input type {} is not supported."
+                           .format(type(source)))
+
+    return result