+++ jwst/resample/resample_utils.py
@@ -6,7 +6,7 @@ from astropy.modeling.models import Scale, AffineTransformation2D
 from astropy.modeling import Model
 from gwcs import WCS, wcstools
 
-from stsci.tools.bitmask import interpret_bit_flags
+from astropy.nddata.bitmask import interpret_bit_flags
 
 from ..assign_wcs.util import wcs_from_footprints, wcs_bbox_from_shape